backend/ltml: connection to main parser
